+++ src/Rules/Required.php
@@ -12,14 +12,23 @@ declare(strict_types = 1);
 namespace Linna\Filter\Rules;
 
 /**
- * Check required.
+ * Check required, value passed must be not null or not 0 length string.
+ *
  */
 class Required implements RuleInterface
 {
     /**
-     * @var array Arguments expected.
+     * @var array Rule properties
      */
-    private $arguments = [];
+    public static $config = [
+        'class' => 'Required',
+        'full_class' => __CLASS__,
+        'alias' => ['required', 'req', 'rq'],
+        'args_count' => 0,
+        'args_type' => [],
+        'has_validate' => true,
+        'has_sanitize' => false
+    ];
 
     /**
      * @var string Error message