+++ CHANGELOG.md
@@ -9,11 +9,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased][v0.2.0](https://github.com/linna/filter/compare/v0.1.0...v0.2.0) - 2018-XX-XX
 
 ### Added
+* Rule aliases
 * New rule and more human readable message system
 * `Linna\Filter->filterOne()` and `Linna\Filter->filterMulti()` methods now return a results object
-* `Linna\Filter\Rules\NumberIntervall` validate is inside or outside a range with operators ><, <>, >=<, <=
+* `Linna\Filter\Rules\NumberIntervall` validate if a number inside or outside a range with operators ><, <>, >=<, <=
 * `Linna\Filter\Rules\NumberCompare` validate and compare numbers with operators <, >, >=, <=, =
 * `Linna\Filter\Rules\StringLenCompare` validate and compare length of a string with operators <, >, >=, <=, =, !=
+* `Linna\Filter\Rules\Str` sanitize strings (convert data type to string)
 
 ### Removed
 * `Linna\Filter\Rules\Between`
@@ -29,9 +31,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * `Linna\Filter\Rules\Date` validate a date
 * `Linna\Filter\Rules\DateCompare` validate dates with operators <, >, >=, <=, =
 * `Linna\Filter\Rules\Email` validate an email
-* `Linna\Filter\Rules\Escape` escape output
+* `Linna\Filter\Rules\Escape` convert special chars to html entities
 * `Linna\Filter\Rules\Max` validate numbers, lower than 
 * `Linna\Filter\Rules\MaxLength` validate strings length, lower than
 * `Linna\Filter\Rules\Min` validate numbers, higher than
 * `Linna\Filter\Rules\MinLength` validate strings length, higher than
-* `Linna\Filter\Rules\Number` sanitize numbers
+* `Linna\Filter\Rules\Number` sanitize numbers (convert data type to int or float)